requirements.txt
numpy==1.24.2
Pillow==9.4.0
matplotlib
C:\ScriptAdminRoot\Execute\main\Install-dlib-main\dlib-19.22.99-cp39-cp39-win_amd64.whl
cmake
face-recognition==1.3.0
deepface
opencv-python